+++ test/unit/pegStabilityModules/PSM.t.sol
@@ -10,13 +10,12 @@ import {ESynth, IEVC} from "../../../src/Synths/ESynth.sol";
 import {TestERC20} from "../../mocks/TestERC20.sol";
 import {EthereumVaultConnector} from "ethereum-vault-connector/EthereumVaultConnector.sol";
 
-
 contract PSMTest is Test {
     // TODO tests where ETH transfers fail
 
-    uint public TO_UNDERLYING_FEE = 30;
-    uint public TO_SYNTH_FEE = 30;
-    uint public BPS_SCALE = 10000;
+    uint256 public TO_UNDERLYING_FEE = 30;
+    uint256 public TO_SYNTH_FEE = 30;
+    uint256 public BPS_SCALE = 10000;
 
     ESynth public synth;
     TestERC20 public underlying;
@@ -28,7 +27,7 @@ contract PSMTest is Test {
     address public owner = makeAddr("owner");
     address public wallet1 = makeAddr("wallet1");
     address public wallet2 = makeAddr("wallet2");
-    
+
     function setUp() public {
         // Deploy EVC
         evc = new EthereumVaultConnector();
@@ -71,7 +70,6 @@ contract PSMTest is Test {
         synth.approve(address(psm), 100e18);
         vm.prank(wallet2);
         synth.approve(address(psm), 100e18);
-
     }
 
     function testConstructor() public {
@@ -81,19 +79,19 @@ contract PSMTest is Test {
     }
 
     function testSwapToUnderlyingGivenIn() public {
-        uint amountIn = 10e18;
-        uint expectedAmountOut = amountIn * (BPS_SCALE - TO_UNDERLYING_FEE) / BPS_SCALE;
+        uint256 amountIn = 10e18;
+        uint256 expectedAmountOut = amountIn * (BPS_SCALE - TO_UNDERLYING_FEE) / BPS_SCALE;
 
-        uint swapperSynthBalanceBefore = synth.balanceOf(wallet1);
-        uint receiverBalanceBefore = underlying.balanceOf(wallet2);
-        uint psmUnderlyingBalanceBefore = underlying.balanceOf(address(psm));
+        uint256 swapperSynthBalanceBefore = synth.balanceOf(wallet1);
+        uint256 receiverBalanceBefore = underlying.balanceOf(wallet2);
+        uint256 psmUnderlyingBalanceBefore = underlying.balanceOf(address(psm));
 
         vm.prank(wallet1);
         psm.swapToUnderlyingGivenIn(amountIn, wallet2);
 
-        uint swapperSynthBalanceAfter = synth.balanceOf(wallet1);
-        uint receiverBalanceAfter = underlying.balanceOf(wallet2);
-        uint psmUnderlyingBalanceAfter = underlying.balanceOf(address(psm));
+        uint256 swapperSynthBalanceAfter = synth.balanceOf(wallet1);
+        uint256 receiverBalanceAfter = underlying.balanceOf(wallet2);
+        uint256 psmUnderlyingBalanceAfter = underlying.balanceOf(address(psm));
 
         assertEq(swapperSynthBalanceAfter, swapperSynthBalanceBefore - amountIn);
         assertEq(receiverBalanceAfter, receiverBalanceBefore + expectedAmountOut);
@@ -101,19 +99,19 @@ contract PSMTest is Test {
     }
 
     function testSwapToUnderlyingGivenOut() public {
-        uint amountOut = 10e18;
-        uint expectedAmountIn = amountOut * BPS_SCALE / (BPS_SCALE - TO_UNDERLYING_FEE);
+        uint256 amountOut = 10e18;
+        uint256 expectedAmountIn = amountOut * BPS_SCALE / (BPS_SCALE - TO_UNDERLYING_FEE);
 
-        uint swapperSynthBalanceBefore = synth.balanceOf(wallet1);
-        uint receiverBalanceBefore = underlying.balanceOf(wallet2);
-        uint psmUnderlyingBalanceBefore = underlying.balanceOf(address(psm));
+        uint256 swapperSynthBalanceBefore = synth.balanceOf(wallet1);
+        uint256 receiverBalanceBefore = underlying.balanceOf(wallet2);
+        uint256 psmUnderlyingBalanceBefore = underlying.balanceOf(address(psm));
 
         vm.prank(wallet1);
         psm.swapToUnderlyingGivenOut(amountOut, wallet2);
 
-        uint swapperSynthBalanceAfter = synth.balanceOf(wallet1);
-        uint receiverBalanceAfter = underlying.balanceOf(wallet2);
-        uint psmUnderlyingBalanceAfter = underlying.balanceOf(address(psm));
+        uint256 swapperSynthBalanceAfter = synth.balanceOf(wallet1);
+        uint256 receiverBalanceAfter = underlying.balanceOf(wallet2);
+        uint256 psmUnderlyingBalanceAfter = underlying.balanceOf(address(psm));
 
         assertEq(swapperSynthBalanceAfter, swapperSynthBalanceBefore - expectedAmountIn);
         assertEq(receiverBalanceAfter, receiverBalanceBefore + amountOut);
@@ -121,19 +119,19 @@ contract PSMTest is Test {
     }
 
     function testSwapToSynthGivenIn() public {
-        uint amountIn = 10e18;
-        uint expectedAmountOut = amountIn * (BPS_SCALE - TO_SYNTH_FEE) / BPS_SCALE;
+        uint256 amountIn = 10e18;
+        uint256 expectedAmountOut = amountIn * (BPS_SCALE - TO_SYNTH_FEE) / BPS_SCALE;
 
-        uint swapperUnderlyingBalanceBefore = underlying.balanceOf(wallet1);
-        uint receiverSynthBalanceBefore = synth.balanceOf(wallet2);
-        uint psmUnderlyingBalanceBefore = underlying.balanceOf(address(psm));
+        uint256 swapperUnderlyingBalanceBefore = underlying.balanceOf(wallet1);
+        uint256 receiverSynthBalanceBefore = synth.balanceOf(wallet2);
+        uint256 psmUnderlyingBalanceBefore = underlying.balanceOf(address(psm));
 
         vm.prank(wallet1);
         psm.swapToSynthGivenIn(amountIn, wallet2);
 
-        uint swapperUnderlyingBalanceAfter = underlying.balanceOf(wallet1);
-        uint receiverSynthBalanceAfter = synth.balanceOf(wallet2);
-        uint psmUnderlyingBalanceAfter = underlying.balanceOf(address(psm));
+        uint256 swapperUnderlyingBalanceAfter = underlying.balanceOf(wallet1);
+        uint256 receiverSynthBalanceAfter = synth.balanceOf(wallet2);
+        uint256 psmUnderlyingBalanceAfter = underlying.balanceOf(address(psm));
 
         assertEq(swapperUnderlyingBalanceAfter, swapperUnderlyingBalanceBefore - amountIn);
         assertEq(receiverSynthBalanceAfter, receiverSynthBalanceBefore + expectedAmountOut);
@@ -141,19 +139,19 @@ contract PSMTest is Test {
     }
 
     function testSwapToSynthGivenOut() public {
-        uint amountOut = 10e18;
-        uint expectedAmountIn = amountOut * BPS_SCALE / (BPS_SCALE - TO_SYNTH_FEE);
+        uint256 amountOut = 10e18;
+        uint256 expectedAmountIn = amountOut * BPS_SCALE / (BPS_SCALE - TO_SYNTH_FEE);
 
-        uint swapperUnderlyingBalanceBefore = underlying.balanceOf(wallet1);
-        uint receiverSynthBalanceBefore = synth.balanceOf(wallet2);
-        uint psmUnderlyingBalanceBefore = underlying.balanceOf(address(psm));
+        uint256 swapperUnderlyingBalanceBefore = underlying.balanceOf(wallet1);
+        uint256 receiverSynthBalanceBefore = synth.balanceOf(wallet2);
+        uint256 psmUnderlyingBalanceBefore = underlying.balanceOf(address(psm));
 
         vm.prank(wallet1);
         psm.swapToSynthGivenOut(amountOut, wallet2);
 
-        uint swapperUnderlyingBalanceAfter = underlying.balanceOf(wallet1);
-        uint receiverSynthBalanceAfter = synth.balanceOf(wallet2);
-        uint psmUnderlyingBalanceAfter = underlying.balanceOf(address(psm));
+        uint256 swapperUnderlyingBalanceAfter = underlying.balanceOf(wallet1);
+        uint256 receiverSynthBalanceAfter = synth.balanceOf(wallet2);
+        uint256 psmUnderlyingBalanceAfter = underlying.balanceOf(address(psm));
 
         assertEq(swapperUnderlyingBalanceAfter, swapperUnderlyingBalanceBefore - expectedAmountIn);
         assertEq(receiverSynthBalanceAfter, receiverSynthBalanceBefore + amountOut);
@@ -162,37 +160,37 @@ contract PSMTest is Test {
 
     // Test quotes
     function testQuoteToUnderlyingGivenIn() public {
-        uint amountIn = 10e18;
-        uint expectedAmountOut = amountIn * (BPS_SCALE - TO_UNDERLYING_FEE) / BPS_SCALE;
+        uint256 amountIn = 10e18;
+        uint256 expectedAmountOut = amountIn * (BPS_SCALE - TO_UNDERLYING_FEE) / BPS_SCALE;
 
-        uint amountOut = psm.quoteToUnderlyingGivenIn(amountIn);
+        uint256 amountOut = psm.quoteToUnderlyingGivenIn(amountIn);
 
         assertEq(amountOut, expectedAmountOut);
     }
 
     function testQuoteToUnderlyingGivenOut() public {
-        uint amountOut = 10e18;
-        uint expectedAmountIn = amountOut * BPS_SCALE / (BPS_SCALE - TO_UNDERLYING_FEE);
+        uint256 amountOut = 10e18;
+        uint256 expectedAmountIn = amountOut * BPS_SCALE / (BPS_SCALE - TO_UNDERLYING_FEE);
 
-        uint amountIn = psm.quoteToUnderlyingGivenOut(amountOut);
+        uint256 amountIn = psm.quoteToUnderlyingGivenOut(amountOut);
 
         assertEq(amountIn, expectedAmountIn);
     }
 
     function testQuoteToSynthGivenIn() public {
-        uint amountIn = 10e18;
-        uint expectedAmountOut = amountIn * (BPS_SCALE - TO_SYNTH_FEE) / BPS_SCALE;
+        uint256 amountIn = 10e18;
+        uint256 expectedAmountOut = amountIn * (BPS_SCALE - TO_SYNTH_FEE) / BPS_SCALE;
 
-        uint amountOut = psm.quoteToSynthGivenIn(amountIn);
+        uint256 amountOut = psm.quoteToSynthGivenIn(amountIn);
 
         assertEq(amountOut, expectedAmountOut);
     }
 
     function testQuoteToSynthGivenOut() public {
-        uint amountOut = 10e18;
-        uint expectedAmountIn = amountOut * BPS_SCALE / (BPS_SCALE - TO_SYNTH_FEE);
+        uint256 amountOut = 10e18;
+        uint256 expectedAmountIn = amountOut * BPS_SCALE / (BPS_SCALE - TO_SYNTH_FEE);
 
-        uint amountIn = psm.quoteToSynthGivenOut(amountOut);
+        uint256 amountIn = psm.quoteToSynthGivenOut(amountOut);
 
         assertEq(amountIn, expectedAmountIn);
     }
